chore(grida_dev): port reftest bake to playwright + tsx (drop standalone package.json)
Rewrites the Chrome-bake script as TypeScript using `@playwright/test`'s
`chromium` (already present in the workspace via @grida/reftest) and
invokes it through `pnpm --filter @grida/reftest exec tsx`.

This matches the pattern used by `.agents/skills/render-reftest/scripts/
refbrowser_render.ts` — the script resolves its dependency through the
monorepo's existing `node_modules` instead of needing its own
package.json/lockfile/node_modules tree alongside the script.

- Replace `reftest_bake_chrome.mjs` (puppeteer + standalone package.json)
  with `reftest_bake_chrome.ts` (playwright via workspace resolution)
- Update `bake.rs` to spawn `pnpm --filter @grida/reftest exec tsx` and
  drop the `ensure_puppeteer_installed` precheck
- Smoke-tested: chromium launches, fixture-discovery + filter work,
  exits cleanly when no fixtures match
